requirements.txt
Django==3.0.5
djangorestframework==3.11.0
psycopg2-binary==2.8.4
django-cors-headers==3.2.1
drf-yasg==1.17.1
python-dotenv==0.12.0
envparse==0.2.0
django-redis==4.11.0
channels==2.4.0
channels-redis==2.4.2

django_nose